Explore merging obstacle map into polygons
Doesn't help with the problem! Dammit!
But it's some neat code, and it may make things easier for
the visibility calculation. Not sure really.

I don't love the viscalc being blackboxed for exactly this reason
(it's broken and I don't have the control to fix it), but it still
really beats writing it myself.